Moved files from model to DAO.
And created a shopping cart class in model package.
