+++ src/DAO/PODAO.java
@@ -1,4 +1,4 @@
-package model;
+package DAO;
 
 import java.sql.*;
 import java.util.HashMap;
