+++ src/DAO/POItemDAO.java
@@ -1,4 +1,4 @@
-package model;
+package DAO;
 
 import java.sql.*;
 import java.util.HashMap;
@@ -8,13 +8,13 @@ import javax.naming.InitialContext;
 import javax.naming.NamingException;
 import javax.sql.DataSource;
 
-import bean.VisitEventBean;
+import bean.POItemBean;
 
-public class VisitEventDAO {
+public class POItemDAO {
 
 	private DataSource ds;
 
-	public VisitEventDAO() throws ClassNotFoundException {
+	public POItemDAO() throws ClassNotFoundException {
 		try {
 			ds = (DataSource) (new InitialContext()).lookup("java:/comp/env/jdbc/EECS");
 		} catch (NamingException e) {
@@ -22,18 +22,19 @@ public class VisitEventDAO {
 		}
 	}
 
-	public Map<String, VisitEventBean> retrieve(String visitDay, String bookID) throws SQLException {
-		String query = "select V.DAY, V.BID, B.EVENTTYPE from VISITEVENT V where V.DAY = " + visitDay + " and V.BID = " + bookID;
-		Map<String, VisitEventBean> rv = new HashMap<String, VisitEventBean>();
+	public Map<String, POItemBean> retrieve(String POID, String bookID) throws SQLException {
+		String query = "select P.ID, P.BID, P.PRICE from POITEM P where P.ID = " + Integer.parseInt(POID) 
+				+ " and P.BID = " + Integer.parseInt(bookID);
+		Map<String, POItemBean> rv = new HashMap<String, POItemBean>();
 		Connection con = this.ds.getConnection();
 		PreparedStatement p = con.prepareStatement(query);
 		ResultSet r = p.executeQuery();
 		while (r.next()) {
-			String day = r.getString("DAY");
+			int id = r.getInt("ID");
 			String bid = r.getString("BID");
-			String eventtype = r.getString("EVENTTYPE");
-			VisitEventBean ve = new VisitEventBean(day, bid, eventtype);
-			rv.put(day, ve);
+			int price = r.getInt("PRICE");
+			POItemBean POItem = new POItemBean(id, bid, price);
+			rv.put(id + "", POItem);
 		}
 		r.close();
 		p.close();
