+++ src/DAO/VisitEventDAO.java
@@ -1,4 +1,4 @@
-package model;
+package DAO;
 
 import java.sql.*;
 import java.util.HashMap;
@@ -8,13 +8,13 @@ import javax.naming.InitialContext;
 import javax.naming.NamingException;
 import javax.sql.DataSource;
 
-import bean.BookBean;
+import bean.VisitEventBean;
 
-public class BookDAO {
+public class VisitEventDAO {
 
 	private DataSource ds;
 
-	public BookDAO() throws ClassNotFoundException {
+	public VisitEventDAO() throws ClassNotFoundException {
 		try {
 			ds = (DataSource) (new InitialContext()).lookup("java:/comp/env/jdbc/EECS");
 		} catch (NamingException e) {
@@ -22,19 +22,18 @@ public class BookDAO {
 		}
 	}
 
-	public Map<String, BookBean> retrieve(String bookID) throws SQLException {
-		String query = "select B.BID, B.TITLE, B.PRICE, B.CATEGORY from BOOK B where B.BID = " + bookID;
-		Map<String, BookBean> rv = new HashMap<String, BookBean>();
+	public Map<String, VisitEventBean> retrieve(String visitDay, String bookID) throws SQLException {
+		String query = "select V.DAY, V.BID, B.EVENTTYPE from VISITEVENT V where V.DAY = " + visitDay + " and V.BID = " + bookID;
+		Map<String, VisitEventBean> rv = new HashMap<String, VisitEventBean>();
 		Connection con = this.ds.getConnection();
 		PreparedStatement p = con.prepareStatement(query);
 		ResultSet r = p.executeQuery();
 		while (r.next()) {
+			String day = r.getString("DAY");
 			String bid = r.getString("BID");
-			String title = r.getString("TITLE");
-			int price = r.getInt("PRICE");
-			String category = r.getString("CATEGORY");
-			BookBean book = new BookBean(bid, title, price, category);
-			rv.put(bid, book);
+			String eventtype = r.getString("EVENTTYPE");
+			VisitEventBean ve = new VisitEventBean(day, bid, eventtype);
+			rv.put(day, ve);
 		}
 		r.close();
 		p.close();
